app/requirements-standalone.txt
django>=2.2.24,<2.3.0
asgiref
autopep8
pycodestyle
pytz
sqlparse>=0.4.2,<0.5.0
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
python-dateutil
pymysql
djangorestframework